Moving the MATLAB Robot arm example
Cleaned up a few other minor library comments as well.  Minor stuff
there.

+++ RobotAsciiCom/examples/ArmScripts/ArmScripts.ino
@@ -65,7 +65,7 @@ void jointAngleCallback(byte jointNumber, int jointAngle) {
   Serial.println();
 }
 
-void gripperCallback(byte gripperDistance) {
+void gripperCallback(int gripperDistance) {
   armServos.setGripperDistance(gripperDistance);
   Serial.println();
   Serial.print("Gripper command to ");
